added crateTrackSelectResult to CrateFilterNode
CrateFilterNode no longer interacts with the db, as this is a job for
CrateStorage. To achive this I added a pointer to the TrackCollection
to CrateFilterNode. To get it there I had to pass it through
SearchQueryParser. The latter no longer uses QSqlDatabase to it's constructor.
It gets a pointer to the TrackCollection from where it get's the database.